backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
apscheduler
python-multipart
email-validator
google-genai